requirements.txt
absl-py==0.5.0
astor==0.7.1
certifi==2018.10.15
Django==2.1.2
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.0
numpy==1.15.2
protobuf==3.6.1
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
Werkzeug==0.14.1
Pillow==5.3.0
django-cors-headers==2.4.0
djangorestframework-gis==0.13
djangorestframework-filters==1.0.0.dev0
